LPS-X Ports fetch utility to TS
Disabled temporarily @liferay/portal/no-global-fetch because a valid import
depends on the `.es.js` extension which is no longer the case. Will need to
fix `eslint-config/.../no-global-fetch.js` before merging
