+++ modules/apps/frontend-js/frontend-js-web/types/src/main/resources/META-INF/resources/liferay/util/fetch.d.ts
@@ -11,34 +11,8 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
-
-const DEFAULT_INIT = {
-	credentials: 'include',
-};
-
 /**
  * Fetches a resource. A thin wrapper around ES6 Fetch API, with standardized
  * default configuration.
- * @param {!string|!Request} resource The URL to the resource, or a Resource
- * object.
- * @param {Object=} init An optional object containing custom configuration.
- * @return {Promise} A Promise that resolves to a Response object.
  */
-
-export default function defaultFetch(resource, init = {}) {
-	const headers = new Headers({'x-csrf-token': Liferay.authToken});
-
-	new Headers(init.headers || {}).forEach((value, key) => {
-		headers.set(key, value);
-	});
-
-	const mergedInit = {
-		...DEFAULT_INIT,
-		...init,
-	};
-
-	mergedInit.headers = headers;
-
-	// eslint-disable-next-line @liferay/portal/no-global-fetch
-	return fetch(resource, mergedInit);
-}
+export default function defaultFetch(resource: RequestInfo, init?: RequestInit): Promise<Response>;
